+++ src/main/java/edu/mit/simile/vicino/distances/BZip2Distance.java
@@ -7,7 +7,7 @@ import org.apache.tools.bzip2.CBZip2OutputStream;
 
 public class BZip2Distance extends PseudoMetricDistance {
 
-    public float d2(String x, String y) {
+    public double d2(String x, String y) {
         String str = x + y;
         float result = 0.0f;
         try {
